+++ src/numvec.hpp
@@ -95,7 +95,7 @@ template<class T>
 std::vector<T> to_vec(numvec<T>& in)
 {
 	std::vector<T> out(in.size());
-	for(int i=0;i<in.size(); ++i)
+	for(size_t i=0;i<in.size(); ++i)
 		out[i] = in[i];
 	return out;
 }
